idk
It is a well known fact that warriors and wizards do not get along, because
one side considers the other side to be a collection of bloodthirsty idiots
who can't walk and think at the same time, while the other side is naturally
suspicious of a body of men who mumble a lot and wear long dresses.  Oh, say
the wizards, if we're going to be like that, then, what about all those
studded collars and oiled muscles down at the Young Men's Pagan Association?
To which the heroes reply, that's a pretty good allegation from a bunch of
wimpsoes who won't go near a woman on account, can you believe it, of their
mystical power being sort of drained out.  Right, say the wizards, that just
about does it, you and your leather posing pouches.  Oh yeah, say the the
heroes, why don't you ...
		-- Terry Pratchett, "The Light Fantastic"