idk
A fellow bought a new car, a Nissan, and was quite happy with his purchase.
He was something of an animist, however, and felt that the car really ought
to have a name.  This presented a problem, as he was not sure if the name
should be masculine or feminine.
	After considerable thought, he settled on an naming the car either
Belchazar or Beaumadine, but remained in a quandry about the final choice.
	"Is a Nissan male or female?" he began asking his friends.  Most of
them looked at him pecularly, mumbled things about urgent appointments, and
went on their way rather quickly.
	He finally broached the question to a lady he knew who held a black
belt in judo.  She thought for a moment and answered "Feminine."
	The swiftness of her response puzzled him. "You're sure of that?" he
asked.
	"Certainly," she replied. "They wouldn't sell very well if they were
masculine."
	"Unhhh...  Well, why not?"
	"Because people want a car with a reputation for going when you want
it to.  And, if Nissan's are female, it's like they say...  `Each Nissan, she
go!'"

	[No, we WON'T explain it; go ask someone who practices an oriental
	martial art.  (Tai Chi Chuan probably doesn't count.)  Ed.]