idk
	There was a mad scientist (a mad... social... scientist) who kidnapped
three colleagues, an engineer, a physicist, and a mathematician, and locked
each of them in seperate cells with plenty of canned food and water but no
can opener.
	A month later, returning, the mad scientist went to the engineer's
cell and found it long empty.  The engineer had constructed a can opener from
pocket trash, used aluminum shavings and dried sugar to make an explosive,
and escaped.
	The physicist had worked out the angle necessary to knock the lids
off the tin cans by throwing them against the wall.  She was developing a good
pitching arm and a new quantum theory.
	The mathematician had stacked the unopened cans into a surprising
solution to the kissing problem; his dessicated corpse was propped calmly
against a wall, and this was inscribed on the floor:
	Theorem: If I can't open these cans, I'll die.
	Proof: assume the opposite...