idk
Several years ago, an international chess tournament was being held in a
swank hotel in New York.  Most of the major stars of the chess world were
there, and after a grueling day of chess, the players and their entourages
retired to the lobby of the hotel for a little refreshment.  In the lobby,
some players got into a heated argument about who was the brightest, the
fastest, and the best chess player in the world.  The argument got quite
loud, as various players claimed that honor.  At that point, a security
guard in the lobby turned to another guard and commented, "If there's
anything I just can't stand, it's chess nuts boasting in an open foyer."