idk
The problem ... is that we have run out of dinosaurs to form oil with.
Scientists working for the Department of Energy have tried to form oil using
other animals; they've piled thousands of tons of sand and Middle Eastern
countries on top of cows, raccoons, haddock, laboratory rats, etc., but so
far all they have managed to do is run up an enormous bulldozer-rental bill
and anger a lot of Middle Eastern persons.  None of the animals turned into
oil, although most of the laboratory rats developed cancer.
		-- Dave Barry, "Postpetroleum Guzzler"